build-sys: Import spec file from Fedora
Will be used by next commits.
The SPEC file will allow to an RPM from sources with a simple
rpmbuild command.
Package files are useful for easy deployment.
This will also be used to provide automatic package build.
This commit just contain a verbatim copy of upstream SPEC file
to make easier to see changes to it and have a start point.

Signed-off-by: Frediano Ziglio <fziglio@redhat.com>
Acked-by: Eduardo Lima (Etrunko) <etrunko@redhat.com>